requirements.txt
django
pytorch
numpy
pandas